requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
beautifulsoup4==4.12.2
matplotlib==3.8.2
seaborn==0.13.0
python-dateutil==2.8.2
uuid==1.30
faker==19.6.2
Pillow>=9.0.0
plotly==5.17.0
pycryptodome==3.19.0
cryptography==41.0.7
urllib3==2.0.7
pytesseract==0.3.10
opencv-python==4.8.1.78
openpyxl==3.1.2
pdf2image==1.16.3
pymupdf==1.23.7
tqdm==4.66.1
psutil==5.9.6
scikit-image==0.22.0
croniter==1.4.1
holidays==0.58
openai==1.100.2
pymysql==1.1.0
SQLAlchemy==2.0.23
mysql-connector-python==8.1.0
xlrd==2.0.1
xlwt==1.3.0
XlsxWriter==3.1.2
pytest==8.3.5
PyYAML==6.0.2


